refactor(upstream): extract pool/probe.rs + pool/registration.rs (lab-kvji.12.7 steps 13-14)
- probe.rs: ensure_probe_task + reprobe_upstream (heartbeat/reconnect engine) +
  disabled_reprobe + observability_source tests. pub(super) FLIP (plan §2.1):
  ensure_probe_task (called from discover.rs) + reprobe_upstream (from ensure.rs).
- registration.rs: register_in_process_service_peers/_list/_with_connector + 2
  in_process_registration tests. InProcessConnector/InProcessRegistration types
  stay in pool.rs.
observability_source test relocated to probe.rs keeping its directory-scan body
(can't include_str a single file). Behavior unchanged. 1423 green.

Bead: lab-kvji.12.7
