+++ crates/lab/src/dispatch/upstream/pool.rs
@@ -21,10 +21,8 @@ use tokio_util::sync::CancellationToken;
 
 use crate::config::UpstreamConfig;
 use crate::oauth::upstream::cache::OauthClientCache;
-use crate::registry::{RegisteredService, ToolRegistry};
+use crate::registry::RegisteredService;
 
-use super::transport::websocket::{jitter_delay, reprobe_backoff};
-use super::types;
 use super::types::{
     UpstreamCapability, UpstreamEntry, UpstreamHealth, UpstreamRuntimeMetadata,
     UpstreamRuntimeOwner, UpstreamTool, UpstreamToolExposureRow,
@@ -41,6 +39,8 @@ mod health;
 mod helpers;
 mod lifecycle;
 mod logging;
+mod probe;
+mod registration;
 #[cfg(test)]
 mod testsupport;
 mod validate;
@@ -51,7 +51,6 @@ pub use helpers::{UpstreamCachedSummary, in_process_upstream_name};
 // descendants. Glob-importing the child's `pub(super)` items keeps existing
 // call sites unchanged while the bodies live in the child modules.
 use connect::*;
-use connect_stdio::*;
 use discover::*;
 use entries::*;
 use helpers::*;
@@ -161,413 +160,6 @@ impl UpstreamPool {
     pub(super) fn with_request_timeout(mut self, timeout: Duration) -> Self {
         self.request_timeout = timeout;
         self
-    }
-
-    fn ensure_probe_task(&self, config: UpstreamConfig) {
-        if config.oauth.is_some() {
-            return;
-        }
-
-        let pool = self.clone();
-        tokio::spawn(async move {
-            let mut tasks = pool.probe_tasks.write().await;
-            if tasks.contains_key(&config.name) {
-                return;
-            }
-            let cancel = CancellationToken::new();
-            tasks.insert(config.name.clone(), cancel.clone());
-            drop(tasks);
-            tracing::info!(
-                surface = "dispatch",
-                service = "upstream.pool",
-                action = "upstream.reprobe",
-                event = "scheduled",
-                operation = "health",
-                upstream = %config.name,
-                transport = upstream_transport(&config),
-                "upstream reprobe scheduled"
-            );
-
-            let mut attempt = 0_u32;
-            loop {
-                let base = reprobe_backoff(attempt);
-                let sleep_for = if attempt == 0 {
-                    types::REPROBE_INTERVAL
-                } else {
-                    jitter_delay(base, stable_jitter_seed(&config.name, attempt))
-                };
-                tracing::debug!(
-                    surface = "dispatch",
-                    service = "upstream.pool",
-                    action = "upstream.reprobe",
-                    event = "sleep",
-                    operation = "health",
-                    upstream = %config.name,
-                    transport = upstream_transport(&config),
-                    attempt,
-                    sleep_ms = sleep_for.as_millis(),
-                    "upstream reprobe sleep scheduled"
-                );
-                tokio::select! {
-                    _ = cancel.cancelled() => {
-                        tracing::info!(
-                            surface = "dispatch",
-                            service = "upstream.pool",
-                            action = "upstream.reprobe",
-                            event = "cancelled",
-                            operation = "health",
-                            upstream = %config.name,
-                            transport = upstream_transport(&config),
-                            attempt,
-                            "upstream reprobe cancelled"
-                        );
-                        break;
-                    },
-                    _ = tokio::time::sleep(sleep_for) => {}
-                }
-
-                let reprobe_started = Instant::now();
-                match pool.reprobe_upstream(&config).await {
-                    Ok(true) => {
-                        tracing::info!(
-                            surface = "dispatch",
-                            service = "upstream.pool",
-                            action = "upstream.reprobe",
-                            event = "finish",
-                            operation = "health",
-                            upstream = %config.name,
-                            transport = upstream_transport(&config),
-                            attempt,
-                            elapsed_ms = reprobe_started.elapsed().as_millis(),
-                            changed = true,
-                            "upstream reprobe succeeded"
-                        );
-                        attempt = 0;
-                    }
-                    Ok(false) => {
-                        tracing::debug!(
-                            surface = "dispatch",
-                            service = "upstream.pool",
-                            action = "upstream.reprobe",
-                            event = "finish",
-                            operation = "health",
-                            upstream = %config.name,
-                            transport = upstream_transport(&config),
-                            attempt,
-                            elapsed_ms = reprobe_started.elapsed().as_millis(),
-                            changed = false,
-                            "upstream reprobe skipped"
-                        );
-                    }
-                    Err(error) => {
-                        let kind = classify_upstream_error(&error.to_string());
-                        attempt = attempt.saturating_add(1);
-                        if auth_error_should_backoff_aggressively(kind) {
-                            attempt = attempt.max(AUTH_FAILURE_REPROBE_ATTEMPT_FLOOR);
-                        }
-                        tracing::warn!(
-                            surface = "dispatch",
-                            service = "upstream.pool",
-                            action = "upstream.reprobe",
-                            event = "error",
-                            operation = "health",
-                            upstream = %config.name,
-                            transport = upstream_transport(&config),
-                            attempt,
-                            elapsed_ms = reprobe_started.elapsed().as_millis(),
-                            kind,
-                            error = %error,
-                            "upstream reprobe failed"
-                        );
-                    }
-                }
-            }
-        });
-    }
-
-    async fn reprobe_upstream(&self, config: &UpstreamConfig) -> anyhow::Result<bool> {
-        let started = Instant::now();
-        tracing::debug!(
-            surface = "dispatch",
-            service = "upstream.pool",
-            action = "upstream.reprobe",
-            event = "start",
-            operation = "health",
-            upstream = %config.name,
-            transport = upstream_transport(config),
-            "upstream reprobe start"
-        );
-        let existing_peer = {
-            let connections = self.connections.read().await;
-            connections
-                .get(&config.name)
-                .map(|connection| connection.peer.clone())
-        };
-
-        if let Some(peer) = existing_peer {
-            match tokio::time::timeout(DISCOVERY_TIMEOUT, peer.list_all_tools()).await {
-                Ok(Ok(tools)) => {
-                    self.replace_catalog_tools(config, tools).await;
-                    self.record_success_for(&config.name, UpstreamCapability::Tools)
-                        .await;
-                    tracing::info!(
-                        surface = "dispatch",
-                        service = "upstream.pool",
-                        action = "upstream.reprobe",
-                        event = "heartbeat.finish",
-                        operation = "health",
-                        upstream = %config.name,
-                        transport = upstream_transport(config),
-                        elapsed_ms = started.elapsed().as_millis(),
-                        "upstream heartbeat succeeded"
-                    );
-                    return Ok(true);
-                }
-                Ok(Err(error)) => {
-                    self.record_failure_for(
-                        &config.name,
-                        UpstreamCapability::Tools,
-                        format!("upstream heartbeat failed: {error}"),
-                    )
-                    .await;
-                    tracing::warn!(
-                        surface = "dispatch",
-                        service = "upstream.pool",
-                        action = "upstream.reprobe",
-                        event = "heartbeat.error",
-                        operation = "health",
-                        upstream = %config.name,
-                        transport = upstream_transport(config),
-                        elapsed_ms = started.elapsed().as_millis(),
-                        kind = "upstream_heartbeat_failed",
-                        error = %error,
-                        "upstream heartbeat failed"
-                    );
-                }
-                Err(_) => {
-                    self.record_failure_for(
-                        &config.name,
-                        UpstreamCapability::Tools,
-                        "upstream heartbeat timed out",
-                    )
-                    .await;
-                    tracing::warn!(
-                        surface = "dispatch",
-                        service = "upstream.pool",
-                        action = "upstream.reprobe",
-                        event = "heartbeat.error",
-                        operation = "health",
-                        upstream = %config.name,
-                        transport = upstream_transport(config),
-                        elapsed_ms = started.elapsed().as_millis(),
-                        kind = "timeout",
-                        timeout_secs = DISCOVERY_TIMEOUT.as_secs(),
-                        "upstream heartbeat timed out"
-                    );
-                }
-            }
-        } else {
-            tracing::warn!(
-                surface = "dispatch",
-                service = "upstream.pool",
-                action = "upstream.reprobe",
-                event = "empty",
-                operation = "health",
-                upstream = %config.name,
-                transport = upstream_transport(config),
-                elapsed_ms = started.elapsed().as_millis(),
-                kind = "upstream_not_connected",
-                "upstream reprobe found no existing connection"
-            );
-        }
-
-        let stale_connection = {
-            let mut connections = self.connections.write().await;
-            connections.remove(&config.name)
-        };
-        if let Some(connection) = stale_connection {
-            connection
-                .shutdown(&config.name, "upstream.reprobe.reconnect")
-                .await;
-        }
-
-        let (conn, tools) = connect_upstream(
-            config,
-            None,
-            self.oauth_client_cache.as_ref(),
-            self.runtime_origin.as_deref(),
-            self.runtime_owner.as_ref(),
-        )
-        .await?;
-        {
-            let mut connections = self.connections.write().await;
-            connections.insert(config.name.clone(), conn);
-        }
-        self.replace_catalog_tools(config, tools).await;
-        self.record_success_for(&config.name, UpstreamCapability::Tools)
-            .await;
-        tracing::info!(
-            surface = "dispatch",
-            service = "upstream.pool",
-            action = "upstream.reprobe",
-            event = "reconnect.finish",
-            operation = "health",
-            upstream = %config.name,
-            transport = upstream_transport(config),
-            elapsed_ms = started.elapsed().as_millis(),
-            "upstream reprobe reconnect succeeded"
-        );
-        Ok(true)
-    }
-
-    pub async fn register_in_process_service_peers(&self, registry: &ToolRegistry) {
-        let services: Vec<RegisteredService> = registry
-            .services()
-            .iter()
-            .filter(|service| !service.actions.is_empty())
-            .cloned()
-            .collect();
-        self.register_in_process_service_list(services).await;
-    }
-
-    async fn register_in_process_service_list(&self, services: Vec<RegisteredService>) {
-        let connector: InProcessConnector = Arc::new(|service| {
-            Box::pin(async move {
-                let upstream_name = in_process_upstream_name(service.name);
-                let entry_name: Arc<str> = Arc::from(upstream_name.as_str());
-                let (conn, tools) = connect_in_process_service_peer(&service).await?;
-                Ok(InProcessRegistration {
-                    connection: Some(conn),
-                    tools,
-                    entry_name,
-                    upstream_name,
-                })
-            })
-        });
-        self.register_in_process_service_list_with_connector(services, connector)
-            .await;
-    }
-
-    async fn register_in_process_service_list_with_connector(
-        &self,
-        services: Vec<RegisteredService>,
-        connector: InProcessConnector,
-    ) {
-        let mut in_process_resource_names = Vec::new();
-        let mut futures = FuturesUnordered::new();
-        let mut failed_count = 0usize;
-        let mut timeout_count = 0usize;
-
-        for service in services {
-            let upstream_name = in_process_upstream_name(service.name);
-            tracing::info!(
-                upstream = %upstream_name,
-                service = service.name,
-                timeout_secs = IN_PROCESS_DISCOVERY_TIMEOUT.as_secs(),
-                "starting in-process peer registration"
-            );
-            let connector = Arc::clone(&connector);
-            futures.push(async move {
-                let service_name = service.name;
-                let result =
-                    tokio::time::timeout(IN_PROCESS_DISCOVERY_TIMEOUT, connector(service)).await;
-                (service_name, upstream_name, result)
-            });
-        }
-
-        while let Some((service_name, upstream_name, result)) = futures.next().await {
-            match result {
-                Ok(Ok(registration)) => {
-                    let mut tool_map = HashMap::new();
-                    let tool_count = registration.tools.len();
-                    for tool in registration.tools {
-                        tool_map.insert(
-                            tool.name.to_string(),
-                            cached_upstream_tool(tool, &registration.entry_name).1,
-                        );
-                    }
-
-                    self.catalog.write().await.insert(
-                        registration.upstream_name.clone(),
-                        healthy_in_process_entry(Arc::clone(&registration.entry_name), tool_map),
-                    );
-                    if let Some(conn) = registration.connection {
-                        self.connections
-                            .write()
-                            .await
-                            .insert(registration.upstream_name.clone(), conn);
-                    }
-                    in_process_resource_names.push(registration.upstream_name.clone());
-                    tracing::info!(
-                        upstream = %registration.entry_name,
-                        service = service_name,
-                        tool_count,
-                        resource_count = 0,
-                        prompt_count = 0,
-                        "in-process peer registration succeeded"
-                    );
-                }
-                Ok(Err(error)) => {
-                    failed_count += 1;
-                    let error_message =
-                        format!("failed to register in-process service peer: {error}");
-                    tracing::warn!(
-                        upstream = %upstream_name,
-                        service = service_name,
-                        error = %error_message,
-                        "in-process peer registration failed"
-                    );
-                    let mut catalog = self.catalog.write().await;
-                    let name: Arc<str> = Arc::from(upstream_name.as_str());
-                    let entry = catalog
-                        .remove(&upstream_name)
-                        .map(|existing| {
-                            failed_in_process_entry_from_existing(existing, error_message.clone())
-                        })
-                        .unwrap_or_else(|| failed_in_process_entry(name, error_message));
-                    catalog.insert(upstream_name, entry);
-                }
-                Err(_) => {
-                    failed_count += 1;
-                    timeout_count += 1;
-                    let error_message = format!(
-                        "in-process peer registration timed out after {}s",
-                        IN_PROCESS_DISCOVERY_TIMEOUT.as_secs()
-                    );
-                    tracing::warn!(
-                        upstream = %upstream_name,
-                        service = service_name,
-                        timeout_secs = IN_PROCESS_DISCOVERY_TIMEOUT.as_secs(),
-                        error = %error_message,
-                        "in-process peer registration timed out"
-                    );
-                    let mut catalog = self.catalog.write().await;
-                    let name: Arc<str> = Arc::from(upstream_name.as_str());
-                    let entry = catalog
-                        .remove(&upstream_name)
-                        .map(|existing| {
-                            failed_in_process_entry_from_existing(existing, error_message.clone())
-                        })
-                        .unwrap_or_else(|| failed_in_process_entry(name, error_message));
-                    catalog.insert(upstream_name, entry);
-                }
-            }
-        }
-
-        if !in_process_resource_names.is_empty() {
-            let mut resource_upstreams = self.resource_upstreams.write().await;
-            resource_upstreams.extend(in_process_resource_names);
-            resource_upstreams.sort_unstable();
-            resource_upstreams.dedup();
-        }
-
-        if failed_count > 0 {
-            tracing::warn!(
-                failed_count,
-                timeout_count,
-                "in-process peer registration completed with degraded services"
-            );
-        }
     }
 
     /// Get all healthy upstream tools.
@@ -1859,29 +1451,15 @@ impl Default for UpstreamPool {
 
 #[cfg(test)]
 mod tests {
+    use super::super::types;
     use super::super::types::ToolExposurePolicy;
     use super::testsupport::*;
     use super::*;
     use crate::mcp::logging::logging_level_rank;
     use crate::mcp::server::LabMcpServer;
+    use crate::registry::ToolRegistry;
     use rmcp::model::{LoggingLevel, ResourceContents};
     use std::sync::atomic::{AtomicU8, Ordering};
-
-    #[tokio::test]
-    async fn disabled_upstream_reprobe_is_inert() {
-        let pool = UpstreamPool::new();
-        let mut config = test_upstream_config();
-        config.enabled = false;
-        config.command = Some("definitely-not-spawned".to_string());
-
-        let result = pool
-            .reprobe_tools_for_upstream(&config)
-            .await
-            .expect("disabled reprobe should not error");
-
-        assert!(!result);
-        assert!(pool.find_tool("anything").await.is_none());
-    }
 
     #[test]
     fn merge_upstream_prompts_is_deterministic() {
@@ -2175,209 +1753,6 @@ mod tests {
 
         assert!(pool.find_tool("search_repos").await.is_some());
         assert!(pool.find_tool("delete_repo").await.is_none());
-    }
-
-    #[tokio::test]
-    async fn in_process_registration_isolates_slow_services_from_fast_services() {
-        use futures::future::BoxFuture;
-        use lab_apis::core::action::ActionSpec;
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        static ACTIONS: &[ActionSpec] = &[ActionSpec {
-            name: "status.read",
-            description: "Read status",
-            destructive: false,
-            returns: "Value",
-            params: &[],
-        }];
-
-        fn dispatch(
-            _action: String,
-            _params: Value,
-        ) -> std::pin::Pin<
-            Box<dyn Future<Output = Result<Value, crate::dispatch::error::ToolError>> + Send>,
-        > {
-            Box::pin(async { Ok(Value::Null) })
-        }
-
-        fn service(name: &'static str) -> RegisteredService {
-            RegisteredService {
-                name,
-                description: "test service",
-                category: "test",
-                kind: crate::registry::RegisteredServiceKind::BuiltInUpstreamApi,
-                status: "available",
-                actions: ACTIONS,
-                dispatch,
-            }
-        }
-
-        let pool = UpstreamPool::new();
-        let fast_seen = Arc::new(AtomicUsize::new(0));
-        let fast_seen_for_connector = Arc::clone(&fast_seen);
-        let connector: InProcessConnector = Arc::new(move |service| {
-            let fast_seen = Arc::clone(&fast_seen_for_connector);
-            let future: BoxFuture<'static, anyhow::Result<InProcessRegistration>> =
-                Box::pin(async move {
-                    if service.name == "slow" {
-                        tokio::time::sleep(Duration::from_millis(100)).await;
-                        anyhow::bail!("slow service failed to start");
-                    }
-
-                    fast_seen.fetch_add(1, Ordering::SeqCst);
-                    let upstream_name: Arc<str> = Arc::from(in_process_upstream_name(service.name));
-                    Ok(InProcessRegistration {
-                        connection: None,
-                        tools: Vec::new(),
-                        entry_name: Arc::clone(&upstream_name),
-                        upstream_name: upstream_name.to_string(),
-                    })
-                });
-            future
-        });
-
-        let registration = tokio::spawn({
-            let pool = pool.clone();
-            async move {
-                pool.register_in_process_service_list_with_connector(
-                    vec![service("slow"), service("fast")],
-                    connector,
-                )
-                .await;
-            }
-        });
-
-        tokio::time::sleep(Duration::from_millis(20)).await;
-        assert_eq!(
-            fast_seen.load(Ordering::SeqCst),
-            1,
-            "fast service should register before slow service finishes"
-        );
-
-        registration.await.expect("registration task");
-        assert_eq!(pool.upstream_count().await, 2);
-    }
-
-    #[tokio::test]
-    async fn failed_in_process_registration_does_not_hide_healthy_peer_tools() {
-        use futures::future::BoxFuture;
-        use lab_apis::core::action::ActionSpec;
-
-        static ACTIONS: &[ActionSpec] = &[ActionSpec {
-            name: "status.read",
-            description: "Read status",
-            destructive: false,
-            returns: "Value",
-            params: &[],
-        }];
-
-        fn dispatch(
-            _action: String,
-            _params: Value,
-        ) -> std::pin::Pin<
-            Box<dyn Future<Output = Result<Value, crate::dispatch::error::ToolError>> + Send>,
-        > {
-            Box::pin(async { Ok(Value::Null) })
-        }
-
-        fn service(name: &'static str) -> RegisteredService {
-            RegisteredService {
-                name,
-                description: "test service",
-                category: "test",
-                kind: crate::registry::RegisteredServiceKind::BuiltInUpstreamApi,
-                status: "available",
-                actions: ACTIONS,
-                dispatch,
-            }
-        }
-
-        let pool = UpstreamPool::new();
-        let connector: InProcessConnector = Arc::new(|service| {
-            let future: BoxFuture<'static, anyhow::Result<InProcessRegistration>> =
-                Box::pin(async move {
-                    if service.name == "bad" {
-                        anyhow::bail!("bad service failed to start");
-                    }
-
-                    let upstream_name: Arc<str> = Arc::from(in_process_upstream_name(service.name));
-                    let tool = rmcp::model::Tool::new(
-                        "status.read",
-                        "Read status",
-                        Arc::new(serde_json::Map::new()),
-                    );
-                    Ok(InProcessRegistration {
-                        connection: None,
-                        tools: vec![tool],
-                        entry_name: Arc::clone(&upstream_name),
-                        upstream_name: upstream_name.to_string(),
-                    })
-                });
-            future
-        });
-
-        pool.register_in_process_service_list_with_connector(
-            vec![service("bad"), service("good")],
-            connector,
-        )
-        .await;
-
-        let good_tools = pool
-            .healthy_tools_for_upstream(&in_process_upstream_name("good"))
-            .await;
-        let bad_tools = pool
-            .healthy_tools_for_upstream(&in_process_upstream_name("bad"))
-            .await;
-
-        assert_eq!(good_tools.len(), 1);
-        assert_eq!(good_tools[0].tool.name.as_ref(), "status.read");
-        assert!(bad_tools.is_empty());
-        assert_eq!(pool.upstream_count().await, 2);
-    }
-
-    #[test]
-    fn observability_source_covers_pool_acquire_reprobe_and_drain_events() {
-        // The pool was split into `pool.rs` + the `pool/` child modules, so the
-        // observability instrumentation now lives across several files. Scan the
-        // whole upstream-pool source tree (pool.rs + every pool/*.rs) so this
-        // guard stays robust as code relocates between modules. A missing string
-        // here means a real dropped-instrumentation regression — never delete an
-        // assertion to make this test pass; add the file the string moved into.
-        let dir = concat!(env!("CARGO_MANIFEST_DIR"), "/src/dispatch/upstream");
-        let mut source =
-            std::fs::read_to_string(format!("{dir}/pool.rs")).expect("read pool.rs source");
-        let pool_dir = format!("{dir}/pool");
-        if let Ok(entries) = std::fs::read_dir(&pool_dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.extension().and_then(|e| e.to_str()) == Some("rs") {
-                    source.push_str(
-                        &std::fs::read_to_string(&path).expect("read pool child module source"),
-                    );
-                }
-            }
-        }
-        for expected in [
-            "action = \"upstream.acquire\"",
-            "elapsed_ms",
-            "pool_size",
-            "connection_count",
-            "action = \"upstream.reprobe\"",
-            "operation = \"health\"",
-            "action = \"upstream.pool.drain\"",
-            "cancelled_probe_count",
-            "kind = \"upstream_pool_empty\"",
-            "kind = \"upstream_not_connected\"",
-            "fn log_upstream_request_start",
-            "fn log_upstream_request_finish",
-            "fn log_upstream_request_error",
-            "action = \"upstream.request\"",
-        ] {
-            assert!(
-                source.contains(expected),
-                "missing upstream pool observability field `{expected}`"
-            );
-        }
     }
 
     #[tokio::test]
